+++ diesel/src/mysql/mod.rs
@@ -11,6 +11,6 @@ mod connection;
 mod query_builder;
 pub mod types;
 
-pub use self::backend::{Mysql, MysqlType};
+pub use self::backend::{Mysql, MysqlType, MysqlTypeMetadata};
 pub use self::connection::MysqlConnection;
 pub use self::query_builder::MysqlQueryBuilder;