+++ CHANGELOG.md
@@ -25,6 +25,11 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 [backend-2-0-0]: http://docs.diesel.rs/diesel/backend/trait.Backend.html
 [raw-value-2-0-0]: http://docs.diesel.rs/diesel/backend/type.RawValue.html
 
+* The type metadata for MySQL has been changed to include sign information. If
+  you are implementing `HasSqlType` for `Mysql` manually, or manipulating a
+  `Mysql::TypeMetadata`, you will need to take the new struct
+  `MysqlTypeMetadata` instead.
+
 ### Fixed
 
 * Many types were incorrectly considered non-aggregate when they should not